script_latin: remove duplicate key in th155 balloons
The 3rd part of this file was just changing some values in order to enhance
a bit where the texts are placed, and the 4th part was some actual code to
have a proper text centering.

When I wrote the 4th part, I didn't notice that I was overriding one of the
values from the 3rd part. But even with that, it worked well enough.

In 35ec3ed, we changed our JSON library, which changed the order in which
the lines get applied. And the value from the 3rd part ended up being applied
*after* the one from the 4th part, which broke the text centering code.