script_latin: th155: change arrows in w15x3
w15x3 is the balloon used when the 2 characters of the same team
say the same thing (for example Joon et Shion). When the balloon was
vertical, there was an arrow pointing to the top character and another
one pointing to the bottom character, so having one to the left and
another one to the right made sense. But for horizontal balloons,
both characters are on the left (or on the right for player 2), so
both arrows should point to the left.