th18: prevent centipede replay desync
In order for this fix to work, the replay must be played from stage 1
When the game performs a stage transition, in normal play it will not
reinitialize all cards, when playing a replay it does. Because of that,
stage transitions in replays reset all state of every card that the
player is holding. This isn't a problem because replays store a snapshot
of the card's state for the start of each stage and whenever a new stage
begins, the game simply restores the snapshot into the cards, except
that the centipede's damage multiplier isn't stored in the replay. THis
leads to the following
- Centipede doesn't reset during stage transitions in normal play. The
  damage multiplier persists between stages
- Centipede does reset during stage transition when playing a replay.
  The damage multiplier resets to 0 during stage transitions

And with that we have another case of commit where writing the commit
message takes longer than actually making the commit

Credit to ACK for creating the fix that I implemented here