th10: force original font for font ID 15
Font ID 15 is used by the hint system. I wasn't able to figure out an
absolutely perfect right and center alignment hack that would work with
any font, so I decided to not do any alignment hacking if the text
output functions are being called by the hint system, and forced the
hint system to use the original font no matter what. A lot of hint files
actually rely on the properties of MS Gothic and the ones that I
recieved from GitHub issues try to create lines that span across the
entire STG frame with dashes, and the dashes in Touhou Biolinum are
simply too small.

fixes #47