th095: fix draw_ctext and draw_rtext not using sprintf
In particular this fixes the "100%%" bug in the help lines.

This is essentially adding a missing sprintf_call/sprintf_rep, with the
catch that sprintf_call had to be integrated into the spell_align
binhacks, as these actually rewrite a big chunk of the draw_[cr]text
functions.

I guess the assumption was that none of the callers pass extra
arguments, so it's safe to not run sprintf at all. That's not the case,
because %% still has to be unescaped to %. Granted, in the original
game, you'll never see a string with %% go through draw_[cr]text, but
that can always change due to translations, or, as in this case, due to
some draw_ltext call being changed to draw_ctext. The moral of the story
is to always replace sprintfs with sprintfs, no matter how stupid they
are. Oh, and to use the satisfying and clarifying NOPs :-)

Thanks to Inasaya Flanderin for reporting this issue.