th155: add proper support for v1.21b
Since v1.19_beta4 and v1.21b share the same exe, we'll replace v1.19_beta4
with v1.21b instead of keeping both.

v1.21b and v1.10 share the same exe, but their dat files are wildly
incompatible, so with the current version detection technology in thcrap,
we can't support both, I need to drop support for v1.10.

And with that, it would feel weird to keep support for lesser versions
like v1.10_steam or v1.19_beta4, so I'm dropping those as well. If you want to
see them for historical purpose, they're still in the git history, and if we
ever have better version detection, we can bring them back.