base_tasofro: th175: add marisa_stage3.pl
This file is the only file in the entire game encoded as UTF-8 without BOM
(well, there is also main.pl, but the encoding of that one doesn't matter since
it's ASCII only). And the game seems to have trouble rendering unicode
characters because of that (I'm not sure how the original game can work despite
this issue).